* adds core file
* adds tests for announcer-setters
